Added div to hero section - styled button
happy with header and nav bar.
need to style the div seperater so it fits the drop down menu or figure
out a way of shifting the page down when the menu is open.
Would like to nest header in a container div will create a branch as I'm
worried I may break everything.